refactor: (user-controller) change name to v1
