+++ src/main/java/connectingstar/tars/habit/controller/HabitController.java
@@ -3,7 +3,7 @@ package connectingstar.tars.habit.controller;
 import connectingstar.tars.common.response.DataResponse;
 import connectingstar.tars.habit.query.RunHabitQueryService;
 import connectingstar.tars.habit.request.param.HabitDailyTrackingRequestParam;
-import connectingstar.tars.habit.response.HabitDailyTrackingResponse;
+import connectingstar.tars.habit.response.HabitDailyTrackingGetResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.MediaType;
@@ -26,16 +26,16 @@ import java.util.List;
 @RequiredArgsConstructor
 public class HabitController {
     private final RunHabitQueryService runHabitQueryService;
-    
+
     /**
      * 날짜를 입력받아 해당 날짜의 습관, 기록, 상태를 조회한다.
      * 홈 페이지 - 캘린더 - 날짜별 습관 수행을 조회할 때 사용한다.
      */
     @GetMapping("/daily-trackings")
-    public ResponseEntity<DataResponse<HabitDailyTrackingResponse>> getDailyTrackingList(
+    public ResponseEntity<DataResponse<HabitDailyTrackingGetResponse>> getDailyTrackingList(
             @ModelAttribute @Valid HabitDailyTrackingRequestParam requestParam
     ) {
-        List<HabitDailyTrackingResponse> responseDto = runHabitQueryService.getDailyTrackingList(requestParam);
+        List<HabitDailyTrackingGetResponse> responseDto = runHabitQueryService.getDailyTrackingList(requestParam);
         return ResponseEntity.ok(new DataResponse(responseDto));
     }
 }
